ScaleListener -> Float

+++ app/src/main/java/me/devsaki/hentoid/adapters/ImagePagerAdapter.kt
@@ -411,7 +411,7 @@ class ImagePagerAdapter(context: Context) :
 
                 val scrollLTR = VIEWER_DIRECTION_LTR == displayParams.direction && isScrollLTR
                 ssiv.setOffsetLeftSide(scrollLTR)
-                ssiv.setScaleListener { s -> onAbsoluteScaleChanged(position, s.toFloat()) }
+                ssiv.setScaleListener { s -> onAbsoluteScaleChanged(position, s) }
             }
 
             // Image layout constraints
